feat: a lot of features
* feat: refresh token

* feat: bcrypt hashing for the refresh token

* feat: enabled cors; fix: validation for email; refactor: cleaned auth service

* fix: validation error messages

* feat: tokens when the user registers

* feat: create powst

* feat: form data accepted in controller

* feat: upload image for creating powst (#2)

* feat: schema for powst

* fix: schema for powst

* feat: get all powsts public endpoint

* feat: get user profile & login user details

* feat: get powst by userID

All the powsts created by the user can
now be fetched using '/powst/user/:userId'
endpoint.

The powsts are populated with the user
name and email as well

* feat: public profile setup

Created the basic setup for the
public profile of any user of the
app. Need to work on the fields
that will be sent

* refactor: last name for user

Changed the name to first and last
name for the user

* feat: update profile pic

Profile pic now be uploaded and edited

* feat: user powst

User powst now being sent using the powsts
array stored in the user schema

Refactored the update user details to send
the latest post after reordering the projects

* feat: blog create and get

Blogs can now be created by authenticated users

The users can also fetch all the blogs written
by one single user

* fix: 4 instead of 3 projects and blogs

* fix: creating a powst saves data in user

* feat: find username availability

API created for checking if the username is
available or not

* refactor: cleaned code

* refactor: using username instead of userId for data

* feat: create work experience for user

Created API to create work experience for the user

* feat: get all user works

Get all the user works based on the username of the user

* feat: default user avatar

When the user creates an account, a default image is assigned to the
profilePic key

This is to have a fallback in case the user does not upload the profile
pic or until the time the profile pic is uploaded

* feat: keywords for powst

Added a new prop to powst schema, keywords. We will use this for the
generation of the algorithm to show relevant powst to relevant user

* feat: like and like powst

Created APIs for liking and unliking the post. The APIs first check if
the powst exists or not, if the user has already liked/not yet liked the
powst and then allows the lini